requirements.txt
Django>=5.0
djangorestframework
Pillow
ffmpeg-python
django-cleanup
